Fixed bokeh hist

+++ app.py
@@ -9,9 +9,11 @@ df = fetch()
 
 
 def create_figure(feat_name, bins):
-    hist, edges = np.histogram(df[feat_name], density=True, bins=bins)
+    hist, edges = np.histogram(df[feat_name].astype('float'), density=True,
+            bins=bins)
     p = figure(title=feat_name, tools='', width=600, height=400)
-    p.quad(top=hist, bottom=0, left=edges[:-1], right=edges[1:], fill_color="navy")
+    p.quad(top=hist, bottom=0, left=edges[:-1], right=edges[1:],
+            fill_color="navy")
     p.xaxis.axis_label = feat_name
     p.yaxis.axis_label = 'Count'
     return p
@@ -22,7 +24,8 @@ def index():
     plot = create_figure('adj_close', 10)
     # Embed plot into HTML via Flask Render
     script, div = components(plot)
-    return render_template('index.html', script=script, div=div, table=to_html(df))
+    return render_template('index.html', script=script, div=div,
+        table=to_html(df))
 
 
 @app.route('/about')
